REFACTOR `TrackSpamTest` to `HoneypotTest`
